+++ src/matrix/cpu/dixxi/MultThreads.h
@@ -23,7 +23,7 @@ namespace cpu
                     T* r = result;
 
                     #pragma omp parallel for
-                    for(size_t i = 0; i < size; i++)
+                    for(int i = 0; i < size; i++)
                     {
                         for(size_t j = 0; j < size; j++)
                         {
